requirements.txt
Django>=3.2.12,<4.0
djangorestframework>=3.13.1,<3.14.0
drf-flex-fields>=0.9.8,<0.9.9
psycopg2>=2.8.6,<2.9
Pillow>=9.1.1,<=9.2.0
drf-spectacular>=0.22.1,<0.23
django-cors-headers>=3.12.0,<3.13.0
django-extensions>=3.2.0,<3.4.0
channels>=3.0.4,<=3.0.5
channels-redis>=3.4.0,<=3.4.1
django-import-export>=3.0.1,<=3.0.2
pandas>=1.4.3,<1.5.0
plotly>=5.9.0,<5.10.0